[new] Interfaces for EC and Temperature provider [new] Implementations for new interfaces [new] FanControl uses new interfaces [chg] Updated to WiX 3.8

+++ StagWare.FanControl/Interfaces/ITemperatureFilter.cs
@@ -7,6 +7,6 @@ namespace StagWare.FanControl
 {
     public interface ITemperatureFilter
     {
-        int FilterTemperature(int temperature);
+        double FilterTemperature(double temperature);
     }
 }
